LayoutTests/platform/chromium-mac/fast/dom/prototype-property-expected.txt

Make sure the 'prototype' property on generated Web IDL interfaces is { DontDelete | ReadOnly }.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tryToDeletePrototype(window.HTMLElement) is false
FAIL tryToSetPrototype(window.HTMLElement, null) should be false. Was true.
FAIL tryToSetPrototype(window.HTMLElement, undefined) should be false. Was true.
FAIL tryToSetPrototype(window.HTMLElement, 1) should be false. Was true.
FAIL tryToSetPrototype(window.HTMLElement, window.Object.prototype) should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
